+++ src/main/java/com/oriun/oriun/Controllers/User_sportsController.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import com.oriun.oriun.Models.User_sportsModel;
 import com.oriun.oriun.Services.User_sportsService;
+import com.oriun.oriun.Services.UserService;
+import com.oriun.oriun.Services.SportService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -20,6 +22,10 @@ public class User_sportsController {
     @Autowired
 
     User_sportsService user_sportsService;
+    @Autowired
+    UserService userService;
+    @Autowired
+    SportService sportService;
 
     @GetMapping("/usersports")
     public List<User_sportsModel> obtenerdeporteUsuarios(@RequestParam("user") String user_name){
@@ -29,24 +35,52 @@ public class User_sportsController {
     @PostMapping("/usersportsreg")
 	@ResponseStatus(HttpStatus.OK)
 	public ResponseEntity registersport(@RequestParam("user") String user_name, @RequestParam("sport") String sport) {
-		User_sportsModel usersport= new User_sportsModel();
-        usersport.setNAME_SPORT(sport);
-        usersport.setUSER_NAME(user_name);
-        user_sportsService.saveUser_sport(usersport);
-		return new ResponseEntity<>(
-			"your user sport register is succesfull ", 
-		HttpStatus.OK);
+        if(userService.existUser(user_name)){
+            if(sportService.existSport(sport)){
+                User_sportsModel usersport= new User_sportsModel();
+                usersport.setNAME_SPORT(sport);
+                usersport.setUSER_NAME(user_name);
+                user_sportsService.saveUser_sport(usersport);
+                return new ResponseEntity<>(
+                        "your user sport register is succesfull ",
+                        HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity<>(
+                        "Deporte no encontrado",
+                        HttpStatus.NOT_FOUND);
+            }
+        }
+        else{
+            return new ResponseEntity<>(
+                    "Usuario no encontrado",
+                    HttpStatus.NOT_FOUND);
+        }
 	}
 
     @DeleteMapping("/usersportsdel")
     @ResponseStatus(HttpStatus.OK)
 	public ResponseEntity deletesport(@RequestParam("user") String user_name, @RequestParam("sport") String sport) {
-		User_sportsModel usersport= new User_sportsModel();
-        usersport.setNAME_SPORT(sport);
-        usersport.setUSER_NAME(user_name);
-        user_sportsService.deleteUserSport(usersport);
-		return new ResponseEntity<>(
-			"your user sport delete is succesfull ", 
-		HttpStatus.OK);
+        if(userService.existUser(user_name)){
+            if(sportService.existSport(sport)){
+                User_sportsModel usersport= new User_sportsModel();
+                usersport.setNAME_SPORT(sport);
+                usersport.setUSER_NAME(user_name);
+                user_sportsService.deleteUserSport(usersport);
+                return new ResponseEntity<>(
+                        "your user sport delete is succesfull ",
+                        HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity<>(
+                        "Deporte no encontrado",
+                        HttpStatus.NOT_FOUND);
+            }
+        }
+        else{
+            return new ResponseEntity<>(
+                    "Usuario no encontrado",
+                    HttpStatus.NOT_FOUND);
+        }
 	}
 }
