+++ src/main/java/com/oriun/oriun/Services/LocationSibuService.java
@@ -1,14 +1,18 @@
 package com.oriun.oriun.Services;
 import java.util.ArrayList;
 import java.util.Optional;
+import java.sql.Blob;
 
 import javax.transaction.Transactional;
 
 import com.oriun.oriun.Models.EventModel;
 import com.oriun.oriun.Models.LocationsibuModel;
 
+import com.oriun.oriun.Repositories.ElementRepository;
 import com.oriun.oriun.Repositories.EventRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import com.oriun.oriun.Repositories.LocationsibuRepository;
 
@@ -19,6 +23,9 @@ public class LocationSibuService {
 
     LocationsibuRepository locationsibuRepository;
 
+    @Autowired
+    ElementRepository elementRepository;
+
     public ArrayList<LocationsibuModel> getlocationsibu(){return (ArrayList<LocationsibuModel>)locationsibuRepository.findAll(); }
 
     public LocationsibuModel savelocationsibu(LocationsibuModel locationSibu){
@@ -27,7 +34,19 @@ public class LocationSibuService {
     public Optional<LocationsibuModel> getlocationsibuByName(String name) {
         return locationsibuRepository.findById(name);
     }
-    public LocationsibuModel updatelocationsibu(String name,LocationsibuModel newlocationsibu) {
+    public ResponseEntity updatelocationsibu(LocationsibuModel newlocationsibu) {
+        String nl=newlocationsibu.getNAME_LOCATION();
+        if(locationsibuRepository.existsById(nl)){
+            boolean op=newlocationsibu.isOPEN();
+            Blob im=newlocationsibu.getIMAGE_LOCATION();
+            locationsibuRepository.updatebyID(nl,op,im);
+            return new ResponseEntity<>("Actualizado",
+                    HttpStatus.OK);
+        }
+        else{
+            return new ResponseEntity<>("Ubicacion antigua no encontrada",
+                    HttpStatus.BAD_REQUEST );
+        }/*
         Optional<LocationsibuModel> oldlocationsibu = locationsibuRepository.findById(name);
         if(oldlocationsibu.isPresent()){
             locationsibuRepository.delete(oldlocationsibu.get());
@@ -35,14 +54,21 @@ public class LocationSibuService {
             return updatedLocationsibu;
         }else{
             return oldlocationsibu.get();
-        }
+        }*/
     }
-    public boolean deleteLocationsibu(LocationsibuModel locationsibu){
-        if(locationsibuRepository.existsById(locationsibu.getNAME_LOCATION())){
-            locationsibuRepository.delete(locationsibu);
-            return true;
+    public ResponseEntity deleteLocationsibu(String nls){
+        if(locationsibuRepository.existsById(nls)){
+            //if(elementRepository.findbyLocation(nls).size()>0){
+                elementRepository.deleteElementsinLSibu(nls);
+            //}
+            locationsibuRepository.deleteById(nls);
+            return new ResponseEntity<>("Ubicacion eliminada",
+                    HttpStatus.OK);
         }
-        return false;
+        else{
+            return new ResponseEntity<>("Ubicacion antigua no existia",
+                    HttpStatus.BAD_REQUEST );
+        }
     }
 
 }
